Split routes.rb in smaller files (webui/api)
To make easier to navigate through the routes files, split it in two
domains:
  - webui (config/routes/webui_routes.rb)
  - api (config/routes/api_routes.rb)
